# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Usuarios",
        "description": "Documentación de la API de Gestión de Usuarios con Swagger en Laravel 11",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "6796c44e1f6079c411abbfed066c5907",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado con éxito"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "string",
                                            "example": "juan@example.com"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIs..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "c4eddaf1a2e990cc90e4326c0d731d41",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener todos los usuarios",
                "operationId": "3ab81e3f9b24b4f36ab08d72349ac632",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener un usuario por ID",
                "operationId": "daaf7a6169ee9348c3fddb07778598ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "59e98fba99e974815e27979dd58e2175",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo Nombre"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "nuevo@email.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado con éxito"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "c30c64b8b7730b65c7714c9af4a6f5fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado con éxito"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user-statistics": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener estadísticas de usuarios registrados",
                "operationId": "fc0fca7c52a9110031e2e96d1eae9669",
                "responses": {
                    "200": {
                        "description": "Estadísticas de usuarios obtenidas con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "usuarios_hoy": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "usuarios_esta_semana": {
                                            "type": "integer",
                                            "example": 30
                                        },
                                        "usuarios_este_mes": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Usuario",
                "description": "Modelo de usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "example": "juan@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T12:30:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa el toquen generado al iniciar sesión",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Usuarios",
            "description": "Usuarios"
        }
    ]
}